SLiM script for random sweepstakes
A SLiM script for simulating a population evolving according to a model of random sweepstakes, where  pairs of 
parents  independently produce juveniles (potential offspring) according to a given heavy-tailed distribution.  The population size is constant.  
Each time the total number of juveniles is at least the population size, so the juveniles are  sorted according to the fitness value and  only the surviving juveniles form the 
next generation of reproducing individuals.  In SLiM only  positive or negative mutations are added,  and we then use  msprime to add  neutral mutations.